refactor and improve altimetry test coverage
